✅ Se termino la pagina dinamica con consumo de API
