+++ package.json
@@ -21,6 +21,7 @@
     "@nextui-org/system": "^2.0.10",
     "@nextui-org/theme": "^2.1.9",
     "@react-navigation/drawer": "^6.6.6",
+    "@react-stately/data": "^3.10.3",
     "@tremor/react": "^3.10.0",
     "framer-motion": "^10.16.4",
     "lightweight-charts": "^4.1.0",
