+++ src/components/icons/dashboard/DeleteIcon.jsx
